backend/requirements.txt
# Django Core
Django==5.0.8
djangorestframework==3.15.2
django-filter==24.2
drf-spectacular==0.27.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# Authentication (Passkeys/WebAuthn) - Alternative Bibliothek
webauthn==1.11.1

# Database
psycopg[binary]==3.2.10

# QR Code Generation
qrcode==7.4.2
Pillow==10.4.0

# HTTP Requests für Print-Agent
requests==2.31.0

# Environment Variables
django-environ==0.11.2

# Testing
pytest==8.2.2
pytest-django==4.8.0

# Code Style
black==24.4.2
isort==5.13.2
flake8==7.0.0

# Development
django-debug-toolbar==4.3.0
